move western calendar class
