+++ lib/zakuro/calculation/base/multi_gengou.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../../era/japan/gengou'
-require_relative '../../era/western'
+require_relative '../../era/western/calendar'
 
 # :nodoc:
 module Zakuro
