+++ lib/zakuro/calculation/monthly/first_day.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../era/western'
+require_relative '../../era/western/calendar'
 
 # :nodoc:
 module Zakuro
