+++ lib/zakuro/era/japan/gengou/validator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../western'
+require_relative '../../western/calendar'
 require_relative './type'
 require 'yaml'
