+++ lib/zakuro/era/japan/reki.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../western'
+require_relative '../western/calendar'
 
 # :nodoc:
 module Zakuro
