+++ lib/zakuro/operation/month/type.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'yaml'
-require_relative '../../era/western'
+require_relative '../../era/western/calendar'
 require_relative './validator'
 
 # :nodoc:
