+++ spec/zakuro/operation/operation_spec.rb
@@ -8,7 +8,7 @@ require File.expand_path('../../../' \
                          __dir__)
 
 require File.expand_path('../../../' \
-                          'lib/zakuro/era/western',
+                          'lib/zakuro/era/western/calendar',
                          __dir__)
 
 # rubocop:disable Metrics/BlockLength
